requirements.txt
PyYaml
easywebdav
icalendar
requests